% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{buckbeak}
\alias{buckbeak}
\title{Get the hippogriff's name}
\usage{
buckbeak()
}
\value{
string. The hippogriff's name
}
\description{
buckbeak
}
\details{
Hagrid's hippogriff name
}
\author{
Insights Engineering <ie@example.com>
}
