% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hagrids_pets}
\alias{hagrids_pets}
\title{Get Hagrid's pets' names}
\usage{
hagrids_pets()
}
\value{
string. The spider and the hippogriff's name
}
\description{
hagrids_pets
}
\details{
Hagrid's pets' names
}
\author{
Insights Engineering <ie@example.com>
}
